refactoring nearest neighbours to use floats instead of ints for distances as this allows more flexibility with different distance measures

+++ image/image-local-features/src/main/java/org/openimaj/feature/local/matcher/FastBasicKeypointMatcher.java
@@ -74,12 +74,12 @@ public class FastBasicKeypointMatcher<T extends Keypoint> extends BasicMatcher<T
 			data[i] = keys1.get(i).ivec;
 		
 		int [][] argmins = new int[keys1.size()][2];
-		int [][] mins = new int[keys1.size()][2];
+		float [][] mins = new float[keys1.size()][2];
 		modelKeypointsKNN.searchKNN(data, 2, argmins, mins);
 		
 		for (int i=0; i<keys1.size(); i++) {
-			int distsq1 = mins[i][0];
-			int distsq2 = mins[i][1];
+			float distsq1 = mins[i][0];
+			float distsq2 = mins[i][1];
 			
 			if (10 * 10 * distsq1 < thresh * thresh * distsq2) {
 				matches.add(new Pair<T>(keys1.get(i), modelKeypoints.get(argmins[i][0])));
